Switched from Nether exploring back to ranch expansion. Found the 0 - 0 - 0 point ; apparantly it's best to put/move the portal there; easier to calculate distance etc in Nether, in order to create more predictable return portals for fast travel. 	to be honest, the Nether does not seem that dangerous if you are 	carefull...
The 0 0 0 point is actually about 13Z into the ocean...  Built a dirt
pier to it, and will reinforce with stone before I build the actual
portal; and also secure the surrounding coast with cobblestonewalls and
torches.

Finished adding the lower south wing, during which apparantly I
connected it to the 'boat house'.